+++ src/client.rs
@@ -1,5 +1,6 @@
 use crate::hash_file::hash_file;
 use crate::sync::Loan;
+use crate::HashAlgorithm;
 use crate::{check_status, write_string, ResumableReadString, ResumableWriteFileBlock};
 use anyhow::anyhow;
 use futures::future::OptionFuture;
@@ -11,51 +12,64 @@ use tokio::io::AsyncWriteExt;
 use tokio::net::TcpStream;
 use tokio_stream::StreamExt;
 
-pub async fn new_process(
-    address: String,
-    secret: String,
-    src_path: PathBuf,
-    dest: PathBuf,
-    force_truncate: bool,
-    workers: u8,
-    block_size: usize,
-) -> Result<(), anyhow::Error> {
-    let mut stream = TcpStream::connect(address).await?;
+pub struct ClientProcessOptions {
+    pub address: String,
+    pub secret: String,
+    pub src_path: PathBuf,
+    pub dest: PathBuf,
+    pub force_truncate: bool,
+    pub workers: u8,
+    pub block_size: usize,
+    pub hash_algorithm: HashAlgorithm,
+}
+
+pub async fn new_process(options: ClientProcessOptions) -> Result<(), anyhow::Error> {
+    let mut stream = TcpStream::connect(options.address).await?;
 
     // first send secret
-    stream.write_u8(secret.len().try_into()?).await?;
-    stream.write_all(secret.as_bytes()).await?;
+    stream.write_u8(options.secret.len().try_into()?).await?;
+    stream.write_all(options.secret.as_bytes()).await?;
     check_status(&mut stream).await?;
 
     // send dest path
     write_string(
         &mut stream,
-        dest.to_str()
+        options
+            .dest
+            .to_str()
             .ok_or(anyhow!("Invalid characters in destination path"))?,
     )
     .await?;
 
     // open source file
-    let mut src = File::open(&src_path)?;
+    let mut src = File::open(&options.src_path)?;
     let src_size = src.seek(SeekFrom::End(0))?;
     src.rewind()?;
 
-    // send block size, file size and force flag
-    stream.write_u64(block_size.try_into()?).await?;
+    // send block size, file size, force flag and hash algorithm
+    stream.write_u64(options.block_size.try_into()?).await?;
     stream.write_u64(src_size).await?;
-    stream.write_u8(if force_truncate { 1 } else { 0 }).await?;
+    stream
+        .write_u8(if options.force_truncate { 1 } else { 0 })
+        .await?;
+    stream.write_u8(options.hash_algorithm.into()).await?;
 
     check_status(&mut stream).await?;
 
     // Start the file hashing
-    let mut hasher = hash_file(&src_path, block_size, workers)?;
+    let mut hasher = hash_file(
+        &options.src_path,
+        options.block_size,
+        options.workers,
+        options.hash_algorithm,
+    )?;
     let mut processing_hash = None;
     let mut hash_comparison_over = false;
 
     // Event loop
     let mut resumable_read_string = ResumableReadString::new();
     let mut block_idx = 0usize;
-    let end_block_idx = (src_size as f32 / block_size as f32).ceil() as usize;
+    let end_block_idx = (src_size as f32 / options.block_size as f32).ceil() as usize;
     let mut blocks_idx_to_send = VecDeque::new();
     let mut resumable_write_block: Option<ResumableWriteFileBlock> = None;
     let (mut stream_rx, mut stream_tx) = tokio::io::split(stream);
@@ -86,7 +100,7 @@ pub async fn new_process(
                 if hash_comparison_over && blocks_idx_to_send.is_empty() {
                     break;
                 } else {
-                    prepare_next_write_block(block_size, &mut blocks_idx_to_send, &mut resumable_write_block).await?;
+                    prepare_next_write_block(options.block_size, &mut blocks_idx_to_send, &mut resumable_write_block).await?;
                 }
             },
 
@@ -100,7 +114,7 @@ pub async fn new_process(
             Some((_, block_data)) = hasher.recv(), if processing_hash.is_some() && !hash_comparison_over => {
                 if block_data.hash != processing_hash.unwrap() {
                     blocks_idx_to_send.push_back((block_idx, block_data.data));
-                    prepare_next_write_block(block_size, &mut blocks_idx_to_send, &mut resumable_write_block).await?;
+                    prepare_next_write_block(options.block_size, &mut blocks_idx_to_send, &mut resumable_write_block).await?;
                 }
                 processing_hash = None;
                 block_idx += 1;