+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+- New flag `--hash` which allow to specify the hash algorithm used for computing blocks checksums.
+
 # 1.0.0 (2024-02-01)
 
 Initial release